requirements/base.txt
Django==1.11.*

# Alliance software utils
-e git+git@gitlab.internal.alliancesoftware.com.au:alliance/alliance-django-utils.git@v/0.2#egg=allianceutils

# AWS interface (esp access to S3 storage)
#boto

# Used for concurrent task processing
celery

# Django authentication tools
django-authtools

# Scheduler extension for celery
django-celery-beat

# Use Django ORM to record celery results
django-celery-results

# HTML editor
#django-ckeditor

# Wraps URL views with a decorator
django-decorator-include

# Useful extensions (eg reset_db, shell_plus, runserver_plus)
# 1.7.5 required for picking up mysql database settings correctly
django-extensions>=1.7.5

# Impersonate users
django-hijack

# Export
#django-import-export

# Default views to requiring login
django-stronghold

# Django webpack bundle
django-webpack-loader

# Django APIs
djangorestframework
djangorestframework-camel-case

# Mysql database support
mysqlclient

# Used in handling data for data analysis
pandas

# Dynamic generation of Word docs
python-docx

# django timezones will use this if present
# do NOT specify a version number: we want pip to not install the pytz but
# instead use patched version from os repo if possible, because ubuntu/redhat/
# debian patchs pytz to reads tzdata from os updates instead of files bundled
# with official pytz which only gets updated with pytz itself
pytz

# better django authentication
rules

# Better python directory utils interface
Unipath

# Test data generation
factory_boy

# temp required to make frontend run
SQLAlchemy
